Move the callbacks into a separate module
The sole purpose of this change is trim down the length of the
Deployment class. The callbacks are more relevant to the validation
process and are the logical component to extract
